refactor: extract weight updators from TrainRayActor
